Created acknowledgment module.
