+++ src/commonworks/info.py
@@ -10,132 +10,6 @@ __version__ = '0.0.0'
 __status__ = 'Development'
 
 
-class Acknowledgement(object):
-    """
-    Represents a CWR Acknowledgement of Transaction (ACK).
-    """
-
-    def __init__(self, group_id, transaction_n, transaction_type, transaction_status, date_creation, date_processing,
-                 title=None, submitter_id=None,
-                 recipient_id=None):
-        self._group_id = group_id
-        self._transaction_n = transaction_n
-        self._transaction_type = transaction_type
-        self._transaction_status = transaction_status
-        self._date_creation = date_creation
-        self._date_processing = date_processing
-        self._title = title
-        self._submitter_id = submitter_id
-        self._recipient_id = recipient_id
-
-    @property
-    def date_creation(self):
-        """
-        Creation Date field.
-
-        The Creation Date of the original file that contained the transaction to which this ACK applies.
-
-        Includes time.
-
-        :return: the date of creation
-        """
-        return self._date_creation
-
-    @property
-    def date_processing(self):
-        """
-        Processing Date field.
-
-        The date this transaction or file was formally processed by the recipient.
-
-        :return: the date in which the transaction or file was processed
-        """
-        return self._date_processing
-
-    @property
-    def group_id(self):
-        """
-        Original Group ID field.
-
-        The Group ID within which the original transaction to which  this ACK applies.  Note that if the ACK is a result
-        of a HDR or TRL record problem, set this field to zeroes.
-
-        :return: the original group id
-        """
-        return self._group_id
-
-    @property
-    def recipient_id(self):
-        """
-        Recipient Creation # field.
-
-        The unique identifier assigned by the recipient to this work. This field is required if the ACK is in response
-        to a transaction and if the transaction status indicates that the recipient has accepted the work.
-
-        :return: the recipient id for this work
-        """
-        return self._recipient_id
-
-    @property
-    def submitter_id(self):
-        """
-        Submitter Creation # field.
-
-        The unique identifier assigned by the original submitter to this work. This field is required if the ACK is in
-        response to a transaction.
-
-        :return: the submitter's id for the work
-        """
-        return self._submitter_id
-
-    @property
-    def title(self):
-        """
-        Creation Title field.
-
-        The creation title as delivered by the submitter (i.e. the title of the musical work or audio visual production,
-        …).  This field is required if the ACK is in response to an NWR or REV transaction.
-
-        :return: the creation title
-        """
-        return self._title
-
-    @property
-    def transaction_n(self):
-        """
-        Original Transaction Sequence # field.
-
-        The Transaction Sequence # of the original transaction to which this ACK applies. Note that if the ACK is a
-        result of a HDR or TRL record problem, set this field to zeroes.
-
-        :return: the original transaction sequence number
-        """
-        return self._transaction_n
-
-    @property
-    def transaction_status(self):
-        """
-        Transaction Status field.
-
-        The current status of this transaction.  Values for this field reside in the Transaction Status Table.
-
-        :return: the transaction status
-        """
-        return self._transaction_status
-
-    @property
-    def transaction_type(self):
-        """
-        Original Transaction Type.
-
-        The Transaction Type of the original transaction to which this ACK applies. Note that if the ACK is a result of
-        a HDR or TRL record problem, set this field to HDR or TRL (whichever is applicable).
-
-        :return: the original transaction type
-        """
-        return self._transaction_type
-
-
 class AdditionalRelatedInfo(object):
     """
     Represents a CWR Additional Related Info (ARI).
