+++ src/commonworks/transaction.py
@@ -35,70 +35,6 @@ __version__ = '0.0.0'
 __status__ = 'Development'
 
 
-class AcknowledgementTransaction(object):
-    """
-    Represents a CWR Acknowledgment Transaction (ACK).
-    """
-
-    def __init__(self, agr=None, nwr=None, rev=None, exc=None, messages=None):
-        self._agr = agr
-        self._nwr = nwr
-        self._rev = rev
-        self._exc = exc
-
-        if messages is None:
-            self._messages = []
-        else:
-            self._messages = messages
-
-    @property
-    def agr(self):
-        """
-        Agreement supporting Work Registration field.
-
-        :return: the agreement transaction affected by this acknowledgement
-        """
-        return self._agr
-
-    @property
-    def exc(self):
-        """
-        Existing Work in Conflict field.
-
-        :return: the existing work in conflict transaction affected by this acknowledgement
-        """
-        return self._exc
-
-    @property
-    def messages(self):
-        """
-        Message field.
-
-        List all messages generated as a result of editing this transaction.
-
-        :return: a list with all the messages
-        """
-        return self._messages
-
-    @property
-    def nwr(self):
-        """
-        New Works Registration field.
-
-        :return: the new works transaction affected by this acknowledgement
-        """
-        return self._nwr
-
-    @property
-    def rev(self):
-        """
-        Revised Registration field.
-
-        :return: the revised registration transaction affected by this acknowledgement
-        """
-        return self._rev
-
-
 class TransactionInterestedParty(object):
     """
     Represents the relationship between an Interested Party and an Agreement
